moves step components to the stepper component folder
