+++ ReactApp/reading-list-app/src/components/add-new-book-stepper/book-summary-step.js
